feat: 工作区重构与文档更新，版本升至 0.0.2
- 前端: 设置/环境/响应/测试等面板拆分组件，工作区视图重构
- 后端: 请求/环境命令与存储 upsert 修复（重命名不再主键冲突）
- 文档: 新增 ARCHITECTURE.md，README 按性能/体积卖点重写，示例图入 docs/imags
- 版本: tauri.conf.json / Cargo.toml / Cargo.lock 同步至 0.0.2

+++ crates/fox-test/src/runner.rs
@@ -20,7 +20,7 @@ pub struct EndpointResult {
     pub path: String,
     pub ok: bool,
     pub status: Option<u16>,
-    pub duration_ms: Option<u64>,
+    pub duration_ms: Option<f64>,
     /// 请求级错误（发送失败 / 配置错误）。
     pub request_error: Option<String>,
     /// 断言明细。
